pyproject.toml
[project]
name = "cstar-ocean"
description = "Computational systems for tracking ocean carbon"
readme = "README.md"
urls = { repository = "https://github.com/CWorthy-ocean/C-Star", homepage = "https://www.cworthy.org" }
authors = [
    { name = "Dafydd Stephenson", email = "dafydd@cworthy.org" },
    { name = "Thomas Nicholas", email = "tom@cworthy.org" },
    { name = "Nora Loose", email = "nora@cworthy.org" },
    { name = "Matt Long", email = "matt@cworthy.org" }
]
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Oceanography"
]
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "python-dateutil>=2.8.2",
    "python-dotenv",
    "PyYAML==6.0.2",
    "pooch>=1.8.1",
    "roms_tools[dask]>=2.4.0"
]
keywords = ["MCDR", "CDR", "ocean carbon", "climate"]

[project.optional-dependencies]
test = [
     "pytest>=7.0",
     "roms_tools[dask]==2.4.0"
     ]
dev =[]

[build-system]
requires = [
    "setuptools>=61.0.0",
    "wheel",
    "setuptools_scm[toml]>=7.0",
    "check-manifest"
]

[tool.setuptools_scm]
fallback_version = "9999"

# Configuration for mypy type checker
[tool.mypy]
ignore_missing_imports = true 
exclude = "properties|asv_bench|docs"

# Configuration for ruff linter and formatter
[tool.ruff]
fix = true
force-exclude=true
extend-exclude = ["docs", "*.ipynb"]

[tool.ruff.format]
exclude = ["docs", "*.ipynb"]

[tool.pre-commit]
config = ".pre-commit-config.yaml"

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*", "examples*", "additional_files*", "externals*"]

[tool.setuptools]
include-package-data = true

[tool.pytest.ini_options]
norecursedirs = [".tox", ".git", "__pycache__", ".#*", "*~"]

# Use addopts to ignore specific files and directories
addopts = [
    "--ignore-glob=cstar/cstar_local_config.py",
    "--ignore-glob=cstar/externals/**",
    "--ignore-glob=examples/**.py",
    "--ignore-glob=*~",
    "--ignore-glob=.#*"
]
